Move some trap stuff to `Traps.cpp`
