+++ soh/soh/Enhancements/randomizer/Traps.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include "3drando/text.hpp"
+#include "libultraship/libultra/types.h"
+
+Text GetIceTrapName(uint16_t id);
